requirements.txt
numpy 
pandas
tqdm
scipy
opencv-python
clean-fid
Pillow 
PyYAML 
tensorboard
torch>=1.6  
torchvision
matplotlib
pytz
Werkzeug 
yacs=0.1.8
psutil
ninja